chore: update test programs for wasi http
