+++ crates/test-programs/wasi-http-tests/Cargo.toml
@@ -5,9 +5,8 @@ readme = "README.md"
 edition = "2021"
 publish = false
 
-[lib]
-crate-type = ["cdylib"]
-
 [dependencies]
 anyhow = { workspace = true }
-wit-bindgen = { workspace = true, features = ["macros", "realloc"] }
+wit-bindgen = { workspace = true, default-features = false, features = [
+    "macros",
+] }
